providers/providerinst: Start of a next-generation provider installer
Our existing provider installer was originally built to work with
releases.hashicorp.com and later retrofitted to talk to the official
Terraform Registry. It also assumes a flat namespace of providers.

We're starting a new one here, copying and adapting code from the old one
as necessary, so that we can build out this new API while retaining all
of the existing functionality and then cut over to this new implementation
in a later step.

As of this commit, the installer only supports querying for available
versions of a given provider, and even then only to the extent needed to
help stub out the service discovery, client request, and error handling
approaches.